fix: don't analyze package dependencies so that they will be installed by PDM Fixes #44
