+++ .github/workflows/ci.yml
@@ -11,7 +11,6 @@ jobs:
     strategy:
       matrix:
         python-version: [3.7, 3.8, 3.9, "3.10", "3.11"]
-        tox-version: [3, 4]
 
     steps:
       - uses: actions/checkout@v3
@@ -21,13 +20,9 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies
         run: |
-          python -m pip install "tox>=${{ matrix.tox-version }}.0.0b0,<${{ matrix.tox-version }}.999" .
+          python -m pip install --upgrade tox .
       - name: Test with tox
         run: |
           tox --version
           pyversion="${{ matrix.python-version }}"
-          toxversion="${{ matrix.tox-version }}"
-          if [[ $toxversion == "3" ]]; then
-            mv pdm.tox3.lock pdm.lock;
-          fi
-          tox -e py${pyversion/./}-tox${toxversion}
+          tox -e py${pyversion/./}
