+++ tox.ini
@@ -5,14 +5,11 @@ ignore =
 max-line-length = 88
 
 [tox]
-envlist = py3{7,8,9,10}-tox{3,4}, lint
+envlist = py3{7,8,9,10,11}, lint
 passenv = LD_PRELOAD
-isolated_build = True
 
 [testenv]
 groups = test
-commands_pre =
-    tox4: pip install tox>=4.0.0b2
 commands =
     tox --version
     test {posargs}
